(fix) Избавляемся от папок
