+++ likhachev.aleksey/T1/base-fucntiuons.cpp
