+++ likhachev.aleksey/T1/base-types.cpp
