--- likhachev.aleksey/T1/classes/classes.hpp
@@ -1,6 +1,0 @@
-#ifndef CLASSES_HPP
-#define CLASSES_HPP
-
-#include "shapes/shapesCommon.hpp"
-
-#endif
