--- likhachev.aleksey/T1/classes/shapes/shapes.hpp
@@ -1,13 +1,0 @@
-#ifndef SHAPES_HPP
-#define SHAPES_HPP
-
-#include "shape.hpp"
-#include "inheritedShapes/parallelogram.hpp"
-#include "inheritedShapes/rectangle.hpp"
-#include "inheritedShapes/ring.hpp"
-
-#include "inheritedShapes/parallelogram.cpp"
-#include "inheritedShapes/rectangle.cpp"
-#include "inheritedShapes/ring.cpp"
-
-#endif
