--- likhachev.aleksey/T1/classes/shapes/shapesCommon.hpp
@@ -1,6 +1,0 @@
-#ifndef SHAPES_COMMON_HPP
-#define SHAPES_COMMON_HPP
-
-#include "../../utils/base-types.hpp"
-
-#endif
