+++ likhachev.aleksey/T1/parallelogram.cpp
