+++ likhachev.aleksey/T1/rectangle.cpp
