+++ likhachev.aleksey/T1/ring.cpp
