+++ likhachev.aleksey/T1/shapes.hpp
@@ -1,0 +1,9 @@
+#ifndef SHAPES_HPP
+#define SHAPES_HPP
+
+#include "shape.hpp"
+#include "parallelogram.hpp"
+#include "rectangle.hpp"
+#include "ring.hpp"
+
+#endif
