+++ likhachev.aleksey/T1/utils.hpp
@@ -5,8 +5,4 @@
 #include "base-fucntiuons.hpp"
 #include "shape-functions.hpp"
 
-#include "base-types.cpp"
-#include "base-fucntiuons.cpp"
-#include "shape-functions.cpp"
-
 #endif
